+++ src/Transaction.cs
@@ -7,8 +7,8 @@ namespace ADALotto.Models
 {
     public class Transaction
     {
-        public float? Id { get; set; }
-        public float? Block { get; set; }
+        public long? Id { get; set; }
+        public long? Block { get; set; }
         public IEnumerable<TransactionMeta>? TxMetadata { get; set; }
         public IEnumerable<TransactionIn>? InTxIns { get; set; }
     }
