+++ src/TransactionIn.cs
@@ -6,7 +6,7 @@ namespace ADALottoModels
 {
     public class TransactionIn
     {
-        public float? TxOutId { get; set; }
+        public long? TxOutId { get; set; }
         public int? TxOutIndex { get; set; }
     }
 }
